logica para escolher o melhor serviço baseado nos status

+++ src/main/kotlin/dto/ServiceHealthStatus.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class ServiceHealthStatus(
     val failing: Boolean,
-    val minResponseTime: Int
+    val minResponseTime: Long
 )
